+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/PeripheralWorksCore.kt
@@ -12,6 +12,7 @@ import site.siredvin.peripheralworks.common.setup.Items
 import site.siredvin.peripheralworks.computercraft.ComputerCraftProxy
 import site.siredvin.peripheralworks.computercraft.StorageProvider
 import site.siredvin.peripheralworks.computercraft.peripherals.PeripheraliumHubPeripheral
+import site.siredvin.peripheralworks.computercraft.peripherals.UltimateSensorPeripheral
 import site.siredvin.peripheralworks.computercraft.peripherals.UniversalScannerPeripheral
 import site.siredvin.peripheralworks.computercraft.plugins.specific.SpecificPluginProvider
 import site.siredvin.peripheralworks.computercraft.plugins.specific.SpecificProtectedPluginProviders
@@ -34,11 +35,14 @@ object PeripheralWorksCore {
                 output.accept(Items.NETHERITE_PERIPHERALIUM_MODEM.get().defaultInstance)
                 output.accept(Blocks.PERIPHERAL_CASING.get().asItem().defaultInstance)
                 output.accept(Blocks.UNIVERSAL_SCANNER.get().asItem().defaultInstance)
+                output.accept(Blocks.ULTIMATE_SENSOR.get().asItem().defaultInstance)
                 PeripheraliumPlatform.getTurtleUpgrade(PeripheraliumHubPeripheral.ID.toString())
                     ?.let { PeripheralWorksPlatform.createTurtlesWithUpgrade(it).forEach(output::accept) }
                 PeripheraliumPlatform.getTurtleUpgrade(PeripheraliumHubPeripheral.NETHERITE_ID.toString())
                     ?.let { PeripheralWorksPlatform.createTurtlesWithUpgrade(it).forEach(output::accept) }
                 PeripheraliumPlatform.getTurtleUpgrade(UniversalScannerPeripheral.UPGRADE_ID.toString())
+                    ?.let { PeripheralWorksPlatform.createTurtlesWithUpgrade(it).forEach(output::accept) }
+                PeripheraliumPlatform.getTurtleUpgrade(UltimateSensorPeripheral.UPGRADE_ID.toString())
                     ?.let { PeripheralWorksPlatform.createTurtlesWithUpgrade(it).forEach(output::accept) }
 
                 PeripheraliumPlatform.getPocketUpgrade(PeripheraliumHubPeripheral.ID.toString())
@@ -46,6 +50,8 @@ object PeripheralWorksCore {
                 PeripheraliumPlatform.getPocketUpgrade(PeripheraliumHubPeripheral.NETHERITE_ID.toString())
                     ?.let { PeripheralWorksPlatform.createPocketsWithUpgrade(it).forEach(output::accept) }
                 PeripheraliumPlatform.getPocketUpgrade(UniversalScannerPeripheral.UPGRADE_ID.toString())
+                    ?.let { PeripheralWorksPlatform.createPocketsWithUpgrade(it).forEach(output::accept) }
+                PeripheraliumPlatform.getPocketUpgrade(UltimateSensorPeripheral.UPGRADE_ID.toString())
                     ?.let { PeripheralWorksPlatform.createPocketsWithUpgrade(it).forEach(output::accept) }
             }
     }